gelka_parser_v2_extracted/requirements.txt
pdfplumber>=0.11.0
Pillow>=10.0.0
python-dateutil>=2.8.2
PyYAML>=6.0.1
